added review-bot
Now that polkadot-fellows/xcm-format#65 has been resolved, review-bot can be implemented.

The following configuration has been provided in the file, but it can be changed if you disagree with the content:

### CI Files
Requires **1 dan 3** to review it and a total of **15 points** to merge.

Some possible combinations are:
- 1 dan6
- 1 dan5 & 1 dan3
- 3 dan3
- 1 dan4 & 1 dan3 & 1 dan2

### Proposals
Requires at least **2 dan 2** to review it and a total of **10 points** to merge.

Some possible combinations are:
- 1 `dan5` && any rank
- 2 `dan3`.
- 5 `dan2`.
- 1 `dan2` and 8 `dan1`.

### General files
Requires at least **2 dan 1** to review it and a total of **4 points** to merge.

Some possible combinations are:
- 1 `dan3` or above.
- 4 `dan1`.
- 2 `dan2`.